feat(webgl): Fix WebGL GAL rendering with proper coordinate setup
- Fix coordinate system for 1:1 pixel mapping (was using nm scale)
- Set worldUnitLength to 1/96 to match native test setup
- Add SetLookAtPoint and SetZoomFactor for proper view transformation
- Fix glDrawBuffers for WebGL 2.0 (array index must match attachment)
- Add legacy_gl_stubs.js for wxWidgets compatibility
- Add isReady() check to prevent race condition in Playwright tests

All 28 GAL test scenarios now render correctly in WebGL.

+++ tests/gal-regression/wasm/webgl/webgl_gal.cpp
@@ -751,7 +751,7 @@ void WEBGL_GAL::BeginDrawing()
 
         m_shader->Use();
         m_shader->SetParameter( ufm_fontTexture, (int) FONT_TEXTURE_UNIT );
-        m_shader->SetParameter( ufm_fontTextureWidth, (int) font_image.width );
+        m_shader->SetParameter( ufm_fontTextureWidth, (float) font_image.width );
         m_shader->Deactivate();
         checkGlError( "setting bitmap font sampler as shader parameter", __FILE__, __LINE__ );
